rev comments
