+++ webapp/channels/src/components/logged_in/logged_in.tsx
@@ -17,7 +17,8 @@ import Constants from 'utils/constants';
 import DesktopApp from 'utils/desktop_api';
 import {isKeyPressed} from 'utils/keyboard';
 import {getBrowserTimezone} from 'utils/timezone';
-import * as UserAgent from 'utils/user_agent';
+import {isAndroid, isIos} from 'utils/user_agent';
+import {doesCookieContainsMMUserId} from 'utils/utils';
 
 declare global {
     interface Window {
@@ -89,9 +90,9 @@ export default class LoggedIn extends React.PureComponent<Props> {
         };
 
         // Device tracking setup
-        if (UserAgent.isIos()) {
+        if (isIos()) {
             document.body.classList.add('ios');
-        } else if (UserAgent.isAndroid()) {
+        } else if (isAndroid()) {
             document.body.classList.add('android');
         }
 
@@ -202,7 +203,7 @@ export default class LoggedIn extends React.PureComponent<Props> {
     private handleBeforeUnload = (): void => {
         // remove the event listener to prevent getting stuck in a loop
         window.removeEventListener('beforeunload', this.handleBeforeUnload);
-        if (document.cookie.indexOf('MMUSERID=') > -1 && this.props.currentChannelId && !this.props.isCurrentChannelManuallyUnread) {
+        if (doesCookieContainsMMUserId() && this.props.currentChannelId && !this.props.isCurrentChannelManuallyUnread) {
             this.props.actions.updateApproximateViewTime(this.props.currentChannelId);
         }
         WebSocketActions.close();
