+++ webapp/channels/src/components/root/index.ts
@@ -17,12 +17,6 @@ import {getTeam} from 'mattermost-redux/selectors/entities/teams';
 import {shouldShowTermsOfService, getCurrentUserId} from 'mattermost-redux/selectors/entities/users';
 
 import {loadRecentlyUsedCustomEmojis, migrateRecentEmojis} from 'actions/emoji_actions';
-import {
-    loadConfigAndMe,
-    registerCustomPostRenderer,
-    handleLoginLogoutSignal,
-    redirectToOnboardingOrDefaultTeam,
-} from 'actions/views/root';
 import {getShowLaunchingWorkspace} from 'selectors/onboarding';
 import {shouldShowAppBar} from 'selectors/plugins';
 import {
@@ -36,6 +30,12 @@ import {initializeProducts} from 'plugins/products';
 
 import type {GlobalState} from 'types/store/index';
 
+import {
+    loadConfigAndMe,
+    registerCustomPostRenderer,
+    handleLoginLogoutSignal,
+    redirectToOnboardingOrDefaultTeam,
+} from './actions';
 import Root from './root';
 
 function mapStateToProps(state: GlobalState) {
